library: Initial support for x86/64
Docker offers an official libary where it's possible to access images
directly without the group name. This means the image is called
`openwrt` instead of `openwrtorg/rootfs`.

To be part of the library it has to be added to the official-images
repository[0] and our repository has to offer a build process only
including a Dockerfile. The easiest way appears to modify our build
scripts to push binary images to a git branch, which is later read and
deployed by the official-libary scripts.

This patch adds the rootfs of the x86/64 architecture and related
buildinfo files. All other files are removed.

A custom shadow file is added to prevent login into root. If enabled,
this would be a security risk as the SSH firewall rules don't apply to
the interface setup by Docker.

Branches should follow the same scheme as version on upstream are
called.

Our CI updater script should force push on the specific branches to keep
things up to date.

[0]: https://github.com/docker-library/official-images/

Signed-off-by: Paul Spooren <mail@aparcar.org>